put assignment in its own file, add scalar ops
