+++ src/circuit_proof/mod.rs
@@ -1,4 +1,5 @@
 #![doc(include = "../docs/circuit-proof-math.md")]
 
+pub mod assignment;
 pub mod circuit;
 pub mod r1cs;
