+++ src/circuit_proof/r1cs.rs
@@ -3,12 +3,13 @@
 
 use rand::{CryptoRng, Rng};
 
+use super::assignment::Assignment;
 use super::circuit::{Circuit, CircuitProof, ProverInput, VerifierInput};
 use curve25519_dalek::scalar::Scalar;
 use errors::R1CSError;
 use generators::{Generators, PedersenGenerators};
 use merlin::Transcript;
-use std::ops::{Add, Div, Mul, Sub, Try};
+use std::ops::Try;
 
 /// The variables used in the `LinearCombination` and `ConstraintSystem` structs.
 #[derive(Clone, Debug)]
@@ -18,96 +19,6 @@ pub enum Variable {
     MultiplierRight(usize),  // low-level variable, right input of multiplication gate
     MultiplierOutput(usize), // low-level variable, output multiplication gate
 }
-
-// The assignment value to a variable, as stored in `ConstraintSystem`.
-// Provers create a `Scalar` assignment, while verifiers create an `R1CSError` assignment.
-#[derive(Copy, Clone, Debug)]
-pub enum Assignment {
-    Value(Scalar),
-    Missing(),
-}
-
-impl Assignment {
-    pub fn new(scalar: Scalar) -> Self {
-        Assignment::Value(scalar)
-    }
-
-    pub fn zero() -> Self {
-        Assignment::Value(Scalar::zero())
-    }
-
-    pub fn from_u64(v: u64) -> Self {
-        Assignment::Value(Scalar::from(v))
-    }
-}
-
-impl Add for Assignment {
-    type Output = Self;
-
-    fn add(self, rhs: Self) -> Self {
-        match (self, rhs) {
-            (Assignment::Value(left), Assignment::Value(right)) => Assignment::Value(left + right),
-            (_, _) => Assignment::Missing(),
-        }
-    }
-}
-
-impl Sub for Assignment {
-    type Output = Self;
-
-    fn sub(self, rhs: Self) -> Self {
-        match (self, rhs) {
-            (Assignment::Value(left), Assignment::Value(right)) => Assignment::Value(left - right),
-            (_, _) => Assignment::Missing(),
-        }
-    }
-}
-
-impl Mul for Assignment {
-    type Output = Self;
-
-    fn mul(self, rhs: Self) -> Self {
-        match (self, rhs) {
-            (Assignment::Value(left), Assignment::Value(right)) => Assignment::Value(left * right),
-            (_, _) => Assignment::Missing(),
-        }
-    }
-}
-
-impl Div for Assignment {
-    type Output = Self;
-
-    fn div(self, rhs: Self) -> Self {
-        match (self, rhs) {
-            (Assignment::Value(left), Assignment::Value(right)) => {
-                Assignment::Value(left * right.invert())
-            }
-            (_, _) => Assignment::Missing(),
-        }
-    }
-}
-
-impl Try for Assignment {
-    type Ok = Scalar;
-    type Error = R1CSError;
-
-    fn into_result(self) -> Result<Self::Ok, Self::Error> {
-        match self {
-            Assignment::Value(val) => Ok(val),
-            Assignment::Missing() => Err(R1CSError::MissingAssignment),
-        }
-    }
-
-    fn from_error(_v: Self::Error) -> Self {
-        Assignment::Missing()
-    }
-
-    fn from_ok(v: Self::Ok) -> Self {
-        Assignment::Value(v)
-    }
-}
-
-// implement From_Iter
 
 /// Represents a linear combination of some variables multiplied with their scalar coefficients,
 /// plus a scalar. `ConstraintSystem` expects all linear combinations to evaluate to zero.
